Fixes flow type scoping issues
This adds a special escope Scope when function, class, and typealias definitions include typeParameters which includes those parameters for reference, but is write-through to it's parent when encountering new variables. This is a little weird, but seems as good as we can while monkey-patching escope.

This also fixes some tests which had invalid flow but were written to expect no errors.

Fixes #123