Avoid errors in eslint 1.10.x when excluding destructured properties.
Despite the improvements from #209 with the pattern visitor, it doesn't
actually iterate over the Property node siblings of the SpreadProperty.
Thus, the "ignore properties left of spread" fix in #95 regressed.

Restoring the "old" path of looping over all properties of an ObjectPattern
that contains a SpreadProperty avoids the regression. It seems inelegant,
but I don't have enough insight into the referencer to propose an alternative.